Reformulate the purpose of `WorkspaceSettings` and update docs
`WorkspaceSettings` have become something else from what it was intended
in the beginning with the addition of `refreshProjectsCommand`. This
commit updates the purpose of workspace settings and clarifies that it's
also a user-facing configuration file.

It is important that we keep the following invariants in bloop now that
we have these semantics:

- Bloop should only override those settings that are managed by it.
- Bloop should detect changes that manually change the build loader
  semantics such as the semanticdb settings and based on that update its
  state.

I've checked bloop checks off these boxes for now, so we should be good
with this repurpose of `workspace.settings.json`.